upload-srt-to-elasticsearch/requirements.txt
setuptools
requests
srt
python-dotenv